# Conflicts:
#	BesTools/Main/LeftSlideViewController/UIView_extra.m
#	BesTools/Main/LeftSlideViewController/Weather/Controllers/WZMainViewController.m
#	BesTools/Main/LeftSlideViewController/Weather/ThirdLibs/JSONModel/JSONModelTransformations/JSONValueTransformer.m

+++ BesTools/Main/LeftSlideViewController/Weather/Controllers/WZMainViewController.h
@@ -14,7 +14,7 @@
 @property(nonatomic,strong) UIPageControl *pageControl;
 @property(nonatomic,strong) NSMutableArray *viewControllers;
 @property(nonatomic,assign) int curPage ;
-@property(nonatomic,assign) NSInteger totalPages ;
+@property(nonatomic,assign) int totalPages ;
 @property(nonatomic,strong) NSMutableArray *citys;
